// apps/client-v2/src/api/generated/models/ItemDTO.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ItemDTO = {
    /**
     * @description The auto-generated id of the item
     * @type integer
    */
    item_id: number;
    /**
     * @description The description of the item
     * @type string
    */
    description: string;
    /**
     * @description The unit price of the item
     * @type number
    */
    item_value: number;
    /**
     * @description Type/Category of the item
     * @type string | undefined
    */
    item_type?: string;
    /**
     * @description Unit of measurement (e.g., m3, gl, un)
     * @type string | undefined
    */
    item_unit?: string;
};